requirements.txt
attrs==22.2.0
colorama==0.4.6
distlib==0.3.6
exceptiongroup==1.1.0
filelock==3.11.0
iniconfig==2.0.0
jsonpickle==3.0.1
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
parse==1.19.0
parse-type==0.6.0
platformdirs==3.2.0
pluggy==1.0.0
pony==0.7.16
PyMySQL==1.0.2
pytest==7.2.1
pytest-bdd==6.1.1
selenium==3.141.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
virtualenv==20.21.0
allure-pytest
